move classes to separate folder
