+++ classes/nameanalysis.py
