+++ classes/orgs_checks.py
@@ -1,4 +1,4 @@
-"""
+""" This file contains functions that are used by exxtract_related_orgs to determine wether the found orgs are true orgs
 Functions:
 
 - keyword_check
